+++ doc/changes.rst
@@ -34,6 +34,9 @@ Release 1.3.1 (in development)
   * :func:`.getPDBCluster` is deprecated for removal in v1.4, use 
     :func:`.listPDBCluster` instead.
 
+  * :func:`.getReservedWords` is deprecated for removal in v1.4, use 
+    :func:`.listReservedWords` instead.
+
 **Bugfix**:
 
   * Fixed a bug in :class:`.HierView` that would cause wrong assignment of